+++ iOS/Scenes/Sharing/EmailGroupSelectionView/EmailGroupSelectionViewModel.swift
@@ -45,6 +45,7 @@ final class EmailGroupSelectionViewModel: ObservableObject {
     @Published private(set) var isChecking = false
     @Published private(set) var isFetchingMore = false
     @Published private(set) var loading = false
+    @Published private(set) var currentUserEmail: String?
     @Published var selectedGroupInfo: GroupInfo?
     @Published var displayType = SuggestionsDisplayType.suggestion
     @Published private(set) var suggestions: [InviteRecommendationType] = []
@@ -74,7 +75,7 @@ final class EmailGroupSelectionViewModel: ObservableObject {
     }
 
     private var filteredCachedGroupInfos: [InviteRecommendationType]? {
-        email.isEmpty ? cachedGroupInfos : cachedGroupInfos?.filter { $0.name.contains(email) }
+        email.isEmpty ? cachedGroupInfos : cachedGroupInfos?.filter { $0.name.lowercased().contains(email) }
     }
 
     init() {
@@ -174,6 +175,10 @@ final class EmailGroupSelectionViewModel: ObservableObject {
 
     func loadData() async {
         loading = true
+
+        if let userData = try? await userManager.getActiveUserData() {
+            currentUserEmail = userData.user.email
+        }
 
         await fetchGroupsInfos()
         async let fetchSuggestions = fetchSuggestions()