+++ iOS/Scenes/Sharing/ManageSharedShareView/ManageSharedShareView.swift
@@ -47,6 +47,7 @@ struct ManageSharedShareView: View {
         }
         .animation(.default, value: viewModel.fetching)
         .task {
+            await viewModel.setUp()
             viewModel.fetchShareInformation(displayFetchingLoader: true)
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity)
@@ -71,7 +72,8 @@ struct ManageSharedShareView: View {
                        "Vaults can’t contain more than 10 users.")
                })
         .sheet(item: $viewModel.selectedGroupInfo) { groupInfo in
-            GroupUsersInformationView(groupInfo: groupInfo, rights: nil)
+            GroupUsersInformationView(groupInfo: groupInfo,
+                                      currentUserEmail: viewModel.currentUserEmail)
                 .presentationDetents([.medium, .large])
                 .presentationDragIndicator(.visible)
         }
@@ -197,11 +199,17 @@ private extension ManageSharedShareView {
         if isVaultSelection, invitee.isPending {
             return true
         }
-
-        return canAdmin &&
-            !invitee.owner &&
-            !viewModel.isCurrentUser(invitee) &&
-            !viewModel.isInCurrentGroup(invitee)
+        if isVaultSelection {
+            return canAdmin &&
+                !invitee.owner &&
+                !viewModel.isCurrentUser(invitee) &&
+                (!viewModel.isInCurrentGroup(invitee) || !invitee.isManager)
+        } else {
+            return canAdmin &&
+                !invitee.owner &&
+                !viewModel.isCurrentUser(invitee) &&
+                (!viewModel.isInCurrentGroup(invitee) || viewModel.currentUserHasDirectMembership())
+        }
     }
 }